build script formatting & optional args to force compile/download
also mac build release lib only

+++ build-deps.sh
@@ -1,9 +1,23 @@
-currentOS="win"
+# Optional args
+    FORCE_DOWNLOAD=false
+    FORCE_COMPILE=false
+
+    ARG1=${1:-"n"}
+    ARG2=${2:-"n"}
+
+    if [ $ARG1 == "-d" ] || [ $ARG2 == "-d" ]; then
+        FORCE_DOWNLOAD=true
+    elif [ $ARG1 == "-c" ] || [ $ARG2 == "-c" ]; then
+        FORCE_COMPILE=true
+    fi
+#
+
 # Get OS type
-    if [[ "$OSTYPE" == "darwin"* ]]; then
+    currentOS="win"
+    if [ "$OSTYPE" == "darwin"* ]; then
         # Mac OSX
         currentOS="mac"
-    elif [[ "$OSTYPE" == "msys" || "$OSTYPE" == "win32" ]]; then
+    elif [ "$OSTYPE" == "msys" ] || [ "$OSTYPE" == "win32" ]; then
         # Lightweight shell and GNU utilities compiled for Windows (part of MinGW), not sure if win32 can even happen?
         currentOS="win"
     else
@@ -12,10 +26,18 @@ currentOS="win"
     fi
 #
 
+echo "OS discovered as $currentOS"
+
 # Download repos
     echo "--------------------------------------------------"
     echo "downloading repos to deps-pre-build"
     echo ""
+
+    declare -i FORCE_DOWNLOAD_TIP=0
+
+    if [ $FORCE_DOWNLOAD == true ] && [ -d "deps-pre-build" ]; then
+        rm -r -f deps-pre-build
+    fi
 
     if [ ! -d "deps-pre-build" ]; then
         mkdir deps-pre-build
@@ -25,60 +47,78 @@ currentOS="win"
 
     if [ ! -d "flucoma-core" ]; then
         git -c advice.detachedHead=false clone --depth 1 -b main https://github.com/flucoma/flucoma-core
+        echo ""
     else
-        echo "flucoma-core already exists, skipping download"
+        FORCE_DOWNLOAD_TIP+=1
     fi
-    echo ""
 
     if [ ! -d "eigen" ]; then
         git -c advice.detachedHead=false clone --depth 1 -b "3.4.0" https://gitlab.com/libeigen/eigen
+        echo ""
     else
-        echo "eigen already exists, skipping download"
+        FORCE_DOWNLOAD_TIP+=1
     fi
-    echo ""
 
     if [ ! -d "hisstools_library" ]; then
         git -c advice.detachedHead=false clone --depth 1 -b header-only https://github.com/alexharker/hisstools_library
+        echo ""
     else
-        echo "hisstools_library already exists, skipping download"
+        FORCE_DOWNLOAD_TIP+=1
     fi
-    echo ""
 
     if [ ! -d "memory" ]; then
         git -c advice.detachedHead=false clone --depth 1 -b main https://github.com/foonathan/memory
+        echo ""
     else
-        echo "memory already exists, skipping download"
+        FORCE_DOWNLOAD_TIP+=1
     fi
-    echo ""
 
     if [ ! -d "spectra" ]; then
         git -c advice.detachedHead=false clone --depth 1 -b "v1.0.1" https://github.com/yixuan/spectra
+        echo ""
     else
-        echo "spectra already exists, skipping download"
+        FORCE_DOWNLOAD_TIP+=1
     fi
-    echo ""
 
     if [ ! -d "json" ]; then
         git -c advice.detachedHead=false clone --depth 1 -b "v3.11.2" https://github.com/nlohmann/json
+        echo ""
     else
-        echo "json already exists, skipping download"
+        FORCE_DOWNLOAD_TIP+=1
     fi
-    echo ""
+
+    if [ $FORCE_DOWNLOAD_TIP > 0 ]; then
+        echo "$FORCE_DOWNLOAD_TIP downloads already exist and were skipped (force this step with -d)"
+        echo ""
+    fi
 #
 
 # Build memory in a new directory
     echo "--------------------------------------------------"
     echo "cmaking foonathan memory libs"
     echo ""
-    if [ -d "compiled-memory" ]; then
+    MEMORY_CMAKE_NEEDED=false
+
+    if [ $FORCE_COMPILE == true ] && [ -d "compiled-memory" ]; then
         rm -r -f compiled-memory
+        cp -r memory/ compiled-memory/
+        MEMORY_CMAKE_NEEDED=true
     fi
-    cp -r memory/ compiled-memory/
+
+    if [ ! -d "compiled-memory" ]; then
+        cp -r memory/ compiled-memory/
+        MEMORY_CMAKE_NEEDED=true
+    fi
+
+    if [ $MEMORY_CMAKE_NEEDED == false ]; then
+        echo "compiled-memory already exists, skipping compilation (force this step with -c)"
+        echo ""
+    fi
 
     #build foonathan_memory libs
     cd compiled-memory
 
-    if [ "$currentOS" == "win" ]; then
+    if [ $MEMORY_CMAKE_NEEDED == true ] && [ "$currentOS" == "win" ]; then
         echo "windows"
         cmake -DFOONATHAN_MEMORY_BUILD_TESTS=OFF -DFOONATHAN_MEMORY_BUILD_TOOLS=OFF -DFOONATHAN_MEMORY_BUILD_EXAMPLES=OFF .
 
@@ -93,16 +133,9 @@ currentOS="win"
         echo ""
 
         cmake --build . --config Release
-    elif [ "$currentOS" == "mac" ]; then
+    elif [ $MEMORY_CMAKE_NEEDED == true ] && [ "$currentOS" == "mac" ]; then
         echo "macos"
-        cmake -DCMAKE_BUILD_TYPE=DEBUG -DFOONATHAN_MEMORY_BUILD_TESTS=OFF -DFOONATHAN_MEMORY_BUILD_TOOLS=OFF -DFOONATHAN_MEMORY_BUILD_EXAMPLES=OFF .
         cmake -DCMAKE_BUILD_TYPE=RELEASE -DFOONATHAN_MEMORY_BUILD_TESTS=OFF -DFOONATHAN_MEMORY_BUILD_TOOLS=OFF -DFOONATHAN_MEMORY_BUILD_EXAMPLES=OFF .
-        
-        echo ""
-        echo "creating debug lib"
-        echo ""
-
-        cmake --build . --config Debug
 
         echo ""
         echo "creating release lib"
@@ -141,12 +174,14 @@ currentOS="win"
     if [ "$currentOS" == "win" ]; then
         cp  deps-pre-build/compiled-memory/src/Debug/*                          libs/
         cp  deps-pre-build/compiled-memory/src/Release/*                        libs/
+    elif [ "$currentOS" == "mac" ]; then
+        cp  deps-pre-build/compiled-memory/src/foonathan_memory-*.a             libs/
     fi
-
+    
     #copy extra compiled foonathan_memory headers
     cp      deps-pre-build/compiled-memory/src/config_impl.hpp                  deps/memory/
     cp      deps-pre-build/compiled-memory/src/container_node_sizes_impl.hpp    deps/memory/
 #
 
-echo ""
+echo "--------------------------------------------------"
 echo "script finished"
